fix: renamed devices/lasers to devices/laser for consistency
