+++ voxel/devices/laser/base.py
