+++ voxel/devices/laser/coherent/obis_ls.py
