+++ voxel/devices/laser/oxxius/oxxius_lcx.py
@@ -2,7 +2,7 @@ from oxxius_laser import LCX
 from serial import Serial
 
 from voxel.descriptors.deliminated_property import DeliminatedProperty
-from voxel.devices.lasers.base import BaseLaser
+from voxel.devices.laser.base import BaseLaser
 
 
 class OxxiusLCXLaser(BaseLaser):
