+++ voxel/devices/laser/vortran/stradus.py
@@ -1,7 +1,7 @@
 from vortran_laser import StradusLaser as StradusVortran, BoolVal
 
 from voxel.descriptors.deliminated_property import DeliminatedProperty
-from voxel.devices.lasers.base import BaseLaser
+from voxel.devices.laser.base import BaseLaser
 
 MODULATION_MODES = {
     'off': {'external_control': BoolVal.OFF, 'digital_modulation': BoolVal.OFF},
